+++ src/main/java/com/digvi/ecommerce/model/OrderItems.java
@@ -27,9 +27,9 @@ public class OrderItems {
 
     private int quantity;
 
-    private Integer mrpPrice;
+    private double mrpPrice;
 
-    private Integer sellingPrice;
+    private double sellingPrice;
 
     private Long userId;
